+++ dds_web/api/__init__.py
@@ -30,7 +30,8 @@ api = flask_restful.Api(api_blueprint, errors=errors)
 
 # New user ############################################################################## New user #
 api.add_resource(admin.InviteUser, "/user/invite", endpoint="invite_user")
-api.add_resource(admin.ConfirmEmail, "/confirm_email/<token>", endpoint="confirm_email")
+api.add_resource(user.ConfirmEmail, "/confirm_email/<token>", endpoint="confirm_email")
+api.add_resource(user.NewUser, "/user/new", endpoint="new_user")
 
 # Login/access ###################################################################### Login/access #
 api.add_resource(user.AuthenticateUser, "/user/auth", endpoint="auth")
