+++ requirements.txt
@@ -9,6 +9,7 @@ Flask-Mail==0.9.1
 Flask-RESTful==0.3.8
 Flask-Security-Too==4.1.0 
 Flask-SQLAlchemy==2.5.1
+Flask-WTF==0.15.1
 immutabledict==2.2.0
 marshmallow-sqlalchemy==0.24.1
 marshmallow==3.13.0 
@@ -19,4 +20,5 @@ PyMySQL==0.10.1
 PyNaCl==1.4.0
 pytest==6.2.4
 pytz==2020.1
+WTForms==2.3.3 
 zstandard==0.15.2
